Update /src/samples itself instead of moving it
