+++ sdk/appconfiguration/azure-data-appconfiguration/src/samples/java/com/azure/data/appconfiguration/AadAuthenticationSample.java
@@ -3,14 +3,21 @@
 
 package com.azure.data.appconfiguration;
 
+import com.azure.core.http.HttpClient;
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.identity.DefaultAzureCredential;
 import com.azure.identity.DefaultAzureCredentialBuilder;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import static com.azure.data.appconfiguration.TestHelper.DISPLAY_NAME_WITH_ARGUMENTS;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 /**
  * Sample demonstrates how to use AAD token to build a configuration client.
  */
-public class AadAuthentication {
+public class AadAuthenticationSample extends ConfigurationClientTestBase {
     /**
      * Sample for how to use AAD token Authentication.
      *
@@ -50,5 +57,32 @@ public class AadAuthentication {
         System.out.printf("[DeleteConfigurationSetting] Key: %s, Value: %s", setting.getKey(), setting.getValue());
 
         System.out.println("End of synchronous sample.");
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.data.appconfiguration.TestHelper#getTestParameters")
+    public void aadAuthenticationSample(HttpClient httpClient, ConfigurationServiceVersion serviceVersion) {
+        // This will manage getting the testing "endpoint" and "tokenCredential" for the test.
+        ConfigurationClient client = setupBuilder(new ConfigurationClientBuilder(), httpClient, serviceVersion, true)
+            .buildClient();
+
+        // Name of the key to add to the configuration service.
+        final String key = testResourceNamer.randomName("key", 16);
+        final String value = "hello world";
+
+        ConfigurationSetting setting = client.setConfigurationSetting(key, null, value);
+        assertEquals(key, setting.getKey());
+        assertNull(setting.getLabel());
+        assertEquals(value, setting.getValue());
+
+        setting = client.getConfigurationSetting(key, null, null);
+        assertEquals(key, setting.getKey());
+        assertNull(setting.getLabel());
+        assertEquals(value, setting.getValue());
+
+        setting = client.deleteConfigurationSetting(key, null);
+        assertEquals(key, setting.getKey());
+        assertNull(setting.getLabel());
+        assertEquals(value, setting.getValue());
     }
 }
