--- sdk/appconfiguration/azure-data-appconfiguration/src/test/java/com/azure/data/appconfiguration/documentation/ConditionalRequestForSettingsPagination.java
@@ -1,70 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.data.appconfiguration.documentation;
-
-import com.azure.core.http.HttpHeaderName;
-import com.azure.core.http.MatchConditions;
-import com.azure.core.http.rest.PagedIterable;
-import com.azure.core.util.Configuration;
-import com.azure.data.appconfiguration.ConfigurationClient;
-import com.azure.data.appconfiguration.ConfigurationClientBuilder;
-import com.azure.data.appconfiguration.models.ConfigurationSetting;
-import com.azure.data.appconfiguration.models.SettingSelector;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-/**
- * Sample demonstrates how to list configuration settings by conditional request.
- * If the ETag of the given setting matches the one in the service, then 304 status code (not modified) with null value
- * returned in the response. Otherwise, a setting with new ETag returned, which is the latest setting retrieved from
- * the service.
- */
-public class ConditionalRequestForSettingsPagination {
-    /**
-     * Runs the sample algorithm and demonstrates how to list configuration settings by conditional request.
-     *
-     * @param args Unused. Arguments to the program.
-     */
-    public static void main(String[] args) {
-        // The connection string value can be obtained by going to your App Configuration instance in the Azure portal
-        // and navigating to the "Access Keys" page under the "Settings" section.
-        String connectionString = Configuration.getGlobalConfiguration().get("AZURE_APPCONFIG_CONNECTION_STRING");
-
-        // Instantiate a client that will be used to call the service.
-        ConfigurationClient client = new ConfigurationClientBuilder()
-                .connectionString(connectionString)
-                .buildClient();
-
-        // list all settings and get their page ETags
-        List<MatchConditions> matchConditionsList = client.listConfigurationSettings(null)
-                .streamByPage()
-                .collect(Collectors.toList())
-                .stream()
-                .map(pagedResponse -> new MatchConditions().setIfNoneMatch(
-                        pagedResponse.getHeaders().getValue(HttpHeaderName.ETAG)))
-                .collect(Collectors.toList());
-
-        PagedIterable<ConfigurationSetting> settings = client.listConfigurationSettings(
-                new SettingSelector().setMatchConditions(matchConditionsList));
-
-        settings.iterableByPage().forEach(pagedResponse -> {
-            int statusCode = pagedResponse.getStatusCode();
-            System.out.println("Status code = " + statusCode);
-            if (statusCode == 304) {
-                System.out.println("Settings have not changed. ");
-                String continuationToken = pagedResponse.getContinuationToken();
-                System.out.println("Continuation Token: " + continuationToken);
-                return;
-            }
-
-            System.out.println("At least one setting in the page has changes. Listing all settings in the page:");
-            System.out.println("new page ETag: " + pagedResponse.getHeaders().getValue(HttpHeaderName.ETAG));
-            pagedResponse.getValue().forEach(setting -> {
-                System.out.println("Key: " + setting.getKey() + ", Value: " + setting.getValue());
-            });
-        });
-    }
-}
-
